Adding pkg/vsphere/disk
The disk package implements a DiskManager which is used to create VMDK
images.  Disks can be created by themselves or from a parent (e.g. a
delta disk).  The API uses VirtualDisk structs as both handles and
state.  It is up to the caller to maintain these structs (cache them,
discover them, etc).  Correctness is implemented by the VirtualDisk
itself by checking if a disk is mounted and/or attached.

The test is very minimal at this point.  It creates a set of virtual
disks which inherit from eachother and attempts to write a portion of a
string to each.  If the leaf of the disks can be used to read the full
string, the test passes.

Still TBD in this package
-Deleting of delta disks by identifying which disk is a leaf
-Creating a whitelist of which disks cannot be mutated (e.g. any cache
or the root disks)

Fixes vmware/vic#11