Updated operations in swagger spec
The docker swagger spec we obtained from go-swagger, which in turn originated
from the docker php project had a lot of mistakes in the operations section.
There were also a lot of missing information as well as incorrect definitions.
This commit cleans up the operations declarations.  It also adds defaults.
Changes are based on Docker Remote API v1.22 spec.