Makefile updates and vicadmin insertion
This makes significant changes to the Makefile, but to general layout, and
to how the targets are resolved. This is intended to allow proper make
behaviour with the generated code portions.
This work was done to support vicadmin, so that's been included as well.
This is vicadmin code from Bonneville, almost entirely unmodified - this
will need significant update in the coming months.