Common iso base and appliance
This creates a common iso base for the appliance and the linux bootstrap.
This needs to be reconnciled with the bootstrap approach but I needed
something immediate and this sufficed.
There is a quirk for some reason extracting the ISO from docker using
stdout:

$ docker run -it test-iso cat /tmp/iso > binary/appliance.iso
$ md5sum binary/appliance.iso
0cf11aa283f3b390e9760fa1c46683ca  binary/appliance.iso
$ docker run -it test-iso md5sum /tmp/iso
b631e1dbb6e1c6178154cb06ba99a5bd  /tmp/iso