integration with port layer
Alters the way components are relaunched so that only those failing or
requiring relaunching are relaunched.

This adds some additional arguments in anticipation of the execution part
of the port layer being added. It makes updates to reflect some of the
changed arguments for the docker server and port layer server in recent
commits.
We need to establish whether it's possible to have both secure/insecure
connections to the docker server now we've switched from the swagger
generated server to the docker-engine project.