whitespace check
The check will be executed via a pre-commit hook and during
the Makefile check.

As a pre-commit hook it will only check the staged files.  From
the Makefile it will check staged and changed files.  If there
is whitespace it will alert and fail -- it will not automatically
correct.

fixes #262