Introduce -resolv to imagec
which returns the metadata of that image that corresponds given reference.
Remove hardcoded values from hackMap and start to fill it via calling
the imagec with new -resolv flag.

Fill required metadata in docker persona and pass it to port-layer.

Change port-layer's exec handler parameters so that they match with tether.

We have now the ability to pull and run images from public or private repos.