Design doc updates (#476)
* configuration mechanism

This adds design doc for the component configuration mechanism (#229) and
reorganised what was there into a more easily consumable structure

* approach commentary

The approach is a basic description, the docker commentary is more
involved. I feel that being entirely open about why we made this decision
is a better approach than attempting to hide those reasons - the reasons
are sane business decisions, but I fear failing to be explicit may result
in people assuming the worst as people are wont to do about big companies
attempting to enter OSS.

* Add note about port layer

Reference to how the portlayer is used, but without going into technical
depth that is present in the architecture docs.

* Documents comms paths for daemon to containers

This adds docs that note the possibilities for communication between
containerVMs and the applianceVM, along with a first pass at the pros &
cons of the different approaches.

* Initial tether design, with appliance interaction

This is a sparse document outlining the responsibilities of the tether and the form of the interaction with the appliance. It's primary use so far is to capture the attach workflow to aid in #439

* Add the possibility of a serial pipe based agent

This is a simpler implementation than the vSocket relay, with a mix of pros & cons from the vsocket relay and the network serial port.

* deployment workflows

This adds initial detail about roles, manifests, and workflow for deploying a VCH.
To do:
- example manifest contents
- links to security doc about signing of manifests and encrypted credential handling
- thoughts on command line args for constructing manifests

* Link to vic-machine

Updates the installation and self-provisioning to link to vic-machine.
I think there's a need for a "usage" document as well that focuses on example user flows without dropping into the technical details.

* Adds diagrams and component naming hints

This converts svgs into png for easy viewing, adds a vic-machine high
level diagram and some vic-machine package name recommendations.


Fixes: #467