Extraconfig update for untagged types (#627)
Adds support for decoding of maps where the value is a pointer to a
struct. Adds tentative guestinfo support as an alternate datasource for decode

Adds support for serializing untagged elements of a nested structure,
thereby allowing the direct use of library structs that do no have vic tags on them.

Adds a guestinfo encoder and error return/checks on the encode paths for all types.

This replaces omitnested with finer granularity controls:
recurse:"depth=n,follow|nofollow"

depth is how many layers to recurse, where a slice of structures is
currently considered to be 2 layers (one for the slice, one for the
structs) - this may need revisting.
follow|nofollow - this controls whether to follow pointers.

Guestinfo encoder/decoder are currently untested due to local VM issues, but not
currently being called anywhere.