Add support for VSAN in the storage layer
VSAN uses namespaces for top-level directories which themselves are VMFS
filesystems.  The top-level directory is a symlink to the VMFS
(represented by a uuid).  This UUID is the stable path to the namespace
and should be used instead of the directory symlink.  This change uses
that and simplifies the datastore URI convetion so we don't have to
compute the URI format for each datastore or fm govmomi API call.  The
`datastore` struct wraps all of the datastore methods and supplies the
right arguments in the appropriate datastore URI format for the given
vsphere datastore.

Tested against VC + VSAN, ESX + VSAN, and ESX.