Use network backing info for debug network (#1181)
* Remove network argument to portlayer

This removes the network argument from the portlayer. It is replaced with
direct access to the extraconfig data. This is specifically to add a debug
network to support our debug NIC. It should be removed anyway eventually,
however the other arguments also need this migration so it was a useful
testbed.

* Update vm test to use generated VM name

One of our tests was using deadbeef for a VM name that gets created on a
datastore. That's caused issues with concurrent testing so is updated to
use the UUID of the guest running the test.
Will likely error if the runner isn't a VM.

Fixes: 1180