Stopping a stopped container throws an error
This fixes two bugs -- both around attempted to stop a stopped
container.  This addresses when a container was created, but
never run then a stop command is issued.  Additionally, when a
started container is stopped and a subsequent stop is issued the
infra layer will throw a vSphere error.

Fixes #1319