Fix docker daemon volumes unit test and persist docker volumes metadata (#1664)
* Fix PL REST API unit test

The test wasn't running and had some copilation issues.  Refactored to
PASS.

Also, pass by ref vs pass by val issues needed some sorting.  This could
probably use a second pass, but it's functional as is.

* Persist volume metadata

We weren't persisting the volume metadata the docker daemon was setting.
We would cache it in the caching layer, but this wouldn't survive
rebootrs.  This change will serialize the metadata and persist it to the
`metadata` directory.  Likewise, on List(), we rehydrate this metadata
when populating the cache.