Abandon inconsistent images and purge images without a manifest file (#1961)
* Abandon images if checksums don't match on pull

This change writes the image compares the computed checksum with the
expected checksum, and moves the disk to the image store if they match.
Otherwise the image's directory is removed.  Subsequent pulls should
complete without an issue.

* Fix image store restart test

The datastore URL was malformed and the test was failing.  Also added
some relevant operation to check the restart actually can find things.

* Write manifest file to signify doneness for images

Write a manifest file (currently empty but free to use for future
stashing of important stuff) for each image after the checksum matches.
This is the last step in a WriteImage so we can use this to know when an
image is complete and consistent.