Move vCenter version check to configs
This removes one step in the installation script for most users
running VC 6.0+. If the user insists on installing the plugin on a
vCenter 5.5 setup then it can be done so by editing variable
IS_VCENTER_5_5 in the configs file. There is a warning that clearly
says 5.5 is not officially supported.