Add Operations to trace contexts [full-ci] (#2016)
* Implements Operations which can be used in place of contexts

The context.Context package is used in our API to share timing and value
state across API boundaries.  One can start a context with a given
deadline, and pass this across APIs, and have the intermediate children
honor the deadline, inherit it, and implement their own.  They also give
cancellation such that the top most (root) context can signal the
children.

This wrapper provides tracing to contexts.  For each Operation created,
a call stack is kept _for that operation_.  It passes API boundaries,
embeds a context so the timout/deadline behavior is maintained, gives
helpful logging utilities to identify unique operations in log messages,
and prints the call stack of the Operation on Err().

* Format pid.unique number for tracing opID in Operations