allow ssh and shell enable via vic-machine (#2555)
Disables ssh and console login by default and allows it to be
re-enabled via the `vic-machine debug` command. This makes use of guest
operations to start sshd, enable shell access for the root user, and set a
password or ssh authorized key.

Adds a robot test for the vic-machine debug command to enable ssh.
This only tests the authorized-key path for now as interactive password is
more awkward.