Lossiness check in attach is optional (#2612)
The lossiness check in attach has a flag controlling it. It's also been
inverted to abide by the requirement that the client always initiates
interactions - this rule avoids the scenario of both sides waiting for
input.
The debug value of the appliance isn't yet plumbed through to this because
the attach connector is started in the interaction handler, not in the
main PL implementation at this time.