Use TLS friendly addresses in inspect output (#2744)
In any vic-machine output that shows the VCH details, attempt to use
the configured certificates to present an address that the host certifcate
will accept.

Selecting between the different certificates that get presented by the
browser is awkward without some form of dis-ambiguation. This uses the
display name and the FQDN/IP if specified as part of the organisation
description if none is supplied by the user.