Fixed anonymous vol issue for postgres in voting app
Our anonymous volume support exposed an issue with postgres and
broke our voting app.  When the modified voting app was created,
it was tested without anonymous volume support.  Our robot
scripts did not have sufficient checks to find this issue.

There is now a workaround in the compose file for postgres and
a check to make sure every container is up and running and not
just the webservers.

Related to #2929.