Add docs for installing/using VIC with Harbor
Provide docs showing how to install VIC and Harbor with self signed certs.
This example workflow illustrate creating a CA cert, server cert, and
server key.  It also illustrates deploying Harbor to a vCenter
installation and configuring it with the server cert and key.  It also
illustrates how to configure a VCH and standard docker to use the CA
cert.  Lastly, it shows the usage of standard docker to tag and push an
image to the Harbor server and usage of the VCH to pull down the image
from Harbor.

Resolves #3102