Add pull-by-digest support for schema 1 images (#5194)
* Add pull-by-digest support for schema 1 images

This commit adds schema 1 support for pulling and removing images
by digest.

  * The crux of the change is type-asserting the parsed image
    reference to check if it is a digest (canonical) type and use
    the digest to fetch the manifest.

  * Per the registry API spec, we verify the manifest payload with
    the input digest shasum for pull-by-digest.

  * In the rmi code, we fetch the digests for the image and untag
    them as well. This is in line with Docker's behavior.

This change supports pull-by-digest for schema 1 only. If a
pull-by-digest results in a schema 2 manifest, the pull operation
will fail with an error. Issue 5187 outlines schema 2 support, after
which we will be able to fully support pull-by-digest.

* Update checks for schema 1 manifest's fields

This change adds checks for schema 1 manifest's FSLayers and History
fields per Docker's code and removes checks for the name and tag
fields.

* Add unit test for getManifestDigest func

This change allows us to also test the case where the image is not
pulled by digest.