Partial fix for issues: #5235, #5245
This fixes a real problem where events go missing. It was seen in multiple/concurrent container create/start.
However, there could be other problems. In particular if an event is not received from VSphere, VIC engine could
still be in an inconsistent state. Where the VM is powered off but the container is still listed as running.