Add toolbox support for listing and killing processes
VIC uses the GuestProcessManager.StartCommand API to run functions within the guest.
Initially, we had only the "kill" function, which does not block.
However, the vSphere API check may block when the target is invalid.
Since the vmx rpc channel is a queue where requests are serialized, such blockage prevents
the toolbox from responding to ping (heartbeat) requests.  In which case, the vmx assumes
toolbox has gone away and responds to requests with VIX_E_TOOLS_NOT_RUNNING (3016).

The toolbox now supports the RPCs to list the state of processes in the guest that were
started via StartCommand.  This state includes the "EndTime" and "ExitCode" fields,
so clients can poll for process completion and still have access to the exit code.

The toolbox functions started by vic-machine (enable-ssh, passwd and API check) no longer block
and vic-machine polls for process completion and exit code via ListProcessesInGuest.

While not used by VIC yet, the toolbox now also supports TerminateProcessInGuest to kill a
process (or function) before is has completed.  Note that VIC still needs to use the "kill"
function (for docker kill/stop) as it takes a signal param, TerminateProcessInGuest does not.

The toolbox also now has support for running actual programs (rather than just functions),
but this feature is not currently enabled within the VCH or container VMs.

Fixes #3417 #3724 #5100