Avoid use of ObjectName for networks
Since vmware/govmomi#734, object.Common.ObjectName() does not work for Network types.
This is due in part to mo.Network having its own "name" field, but ObjectName uses
the mo.ManagedEntity name field.  This can be fixed in govmomi, but for now making
this change in VIC behaves as the previous version did (using base of InventoryPath)
and saves a round trip for each network.