Initial support for custom kernel and tdnf
Outstanding:
* allow each config to supply it's own kernel parameters in isolinux.cfg

Introduces an isos/base/repos directory that holds the
configuration/specification details for a particular variant, including
package mangaer configuration, linux kernel version, and distro packages
required.

This change allows for selection of a specific kernel for the bootstrap
iso and adds initial Photon 2.0 support.

Switches to use udevd or systemd-udevd based based on the init subsystem
type.

Adds a basic breakout of the live init process after bootstrap boot and
makes the use of entropy daemon and iptables a little more generalized.

yum does not support rich dependencies so we will need to move to tdnf as
a package manager during the build in order to adopt photon 2.0 reliably.

Adds per repo configuration of iptables and entropy daemon to allow
customization based on package availability and kernel support.

Switches to building the ISOs in containers, based off either Photon2.0 or
Centos. This is done to get "native" tdnf or yum package managers -
necessary to work around quirks of the two tools.

This also adds convenience targets for the specific vic-machine variants
given a lot of our users use windows and it makes providing build
instructions simpler.