Fix 6-05-Create-Validation tests for vCenter (#6479)
This commit modifies several tests in the 6-05-Create-Validation
suite to run on vCenter. We now specify the image datastore and
the bridge and public networks in VCH create on VC since there is
no default datastore and the networks are needed. The extra params
are specified for VC so that VCH creation succeeds and the
original intent of the tests is satisfied.