Ensure unique container names during create (#6561)
This commit modifies the container create logic in the docker persona
to ensure that containers have unique names, specially during a
parallel container create operation with the same name.

Earlier, the duplicate name check took place at the start of the
create operation, but the name wasn't tracked in the persona cache
until the create operation had succeeded in the portlayer. This made
it possible for multiple containers to have the same name if created
in parallel. Now, the container create function uses the ReserveName
function in the persona's container cache once the container object
representation has been created in the persona. If the container
fails to be created in the portlayer, the name is released.

Fixes #6261