Implement API handler for VCH deletion
Introduce a pair of handlers for deleting VCHs within a vSphere target
or datacenter. By default, deletion includes the VCH and powered off
containers. Deletion may optionally include powered on containers
and/or volume stores as well.

If any containers remain, the VCH is not deleted. If the VCH is not
deleted, the response includes a non-2xx status code.

Define a suite of end-to-end tests which verify the intended deletion
behavior. End-to-end tests do not attempt to verify the behavior of
concurrent operations.